+++ core/src/main/java/tc/oc/pgm/api/integration/FriendIntegration.java
@@ -13,13 +13,4 @@ public interface FriendIntegration {
    * @return true if friends, false if not
    */
   boolean isFriend(Player a, Player b);
-
-  static FriendIntegration Noop = new NoopFriendIntegration();
-
-  public class NoopFriendIntegration implements FriendIntegration {
-    @Override
-    public boolean isFriend(Player a, Player b) {
-      return false;
-    }
-  }
 }
