+++ core/src/main/java/tc/oc/pgm/api/integration/NickIntegration.java
@@ -13,13 +13,4 @@ public interface NickIntegration {
    */
   @Nullable
   String getNick(Player player);
-
-  static NickIntegration Noop = new NoopNickIntegration();
-
-  public class NoopNickIntegration implements NickIntegration {
-    @Override
-    public @Nullable String getNick(Player player) {
-      return null;
-    }
-  }
 }
