+++ core/src/main/java/tc/oc/pgm/api/integration/PunishmentIntegration.java
@@ -21,19 +21,4 @@ public interface PunishmentIntegration {
    */
   @Nullable
   String getMuteReason(Player player);
-
-  static PunishmentIntegration Noop = new NoopPunishmentIntegration();
-
-  public class NoopPunishmentIntegration implements PunishmentIntegration {
-
-    @Override
-    public boolean isMuted(Player player) {
-      return false;
-    }
-
-    @Override
-    public @Nullable String getMuteReason(Player player) {
-      return null;
-    }
-  }
 }
